+++ osquery/singleton.py
@@ -1,0 +1,20 @@
+"""LICENSE file in the root directory of this source tree. An additional grant
+of patent rights can be found in the PATENTS file in the same directory.
+"""
+
+# pylint: disable=too-few-public-methods
+
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
+
+class Singleton(object):
+    """A simple singleton base class"""
+    _instance = None
+
+    def __new__(cls, *args, **kwargs):
+        """Override __new__ to implement custom instantiation"""
+        if not cls._instance:
+            cls._instance = super(Singleton, cls).__new__(cls, *args, **kwargs)
+        return cls._instance
